# contracts/contract3-api.openapi.yaml
openapi: 3.1.0
info:
  title: DarkPulse Investigator API
  version: 1.0.0
paths:
...
  /watchlists:
    get:
      tags:
      - Watchlists
      summary: List Watchlists
      operationId: list_watchlists_api_v1_watchlists_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistListResponse'
      security:
      - HTTPBearer: []
    post:
      tags:
      - Watchlists
      summary: Create Watchlist
      operationId: create_watchlist_api_v1_watchlists_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /watchlists/{watchlist_id}:
    put:
      tags:
      - Watchlists
      summary: Update Watchlist
      operationId: update_watchlist_api_v1_watchlists__watchlist_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: watchlist_id
        in: path
        required: true
        schema:
          type: string
          title: Watchlist Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Watchlists
      summary: Delete Watchlist
      operationId: delete_watchlist_api_v1_watchlists__watchlist_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: watchlist_id
        in: path
        required: true
        schema:
          type: string
          title: Watchlist Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiEnvelope:
      properties:
        data:
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/Pagination'
          - type: 'null'
        meta:
          additionalProperties: true
          type: object
          title: Meta
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
          title: Errors
      type: object
      title: ApiEnvelope
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
      type: object
      required:
      - code
      - message
      title: ErrorDetail
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Pagination:
      properties:
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        limit:
          type: integer
          title: Limit
          default: 50
        total:
          type: integer
          title: Total
          default: 0
      type: object
      title: Pagination
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WatchlistCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        terms:
          items:
            type: string
          type: array
          maxItems: 1000
          minItems: 1
          title: Terms
        notify:
          type: boolean
          title: Notify
          default: true
      type: object
      required:
      - name
      - terms
      title: WatchlistCreate
    WatchlistListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/WatchlistResponse'
          type: array
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/Pagination'
          - type: 'null'
        meta:
          additionalProperties: true
          type: object
          title: Meta
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
          title: Errors
      type: object
      title: WatchlistListResponse
    WatchlistResponse:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        terms:
          items:
            type: string
          type: array
          maxItems: 1000
          minItems: 1
          title: Terms
        notify:
          type: boolean
          title: Notify
          default: true
        enabled:
          type: boolean
          title: Enabled
          default: true
        id:
          type: string
          title: Id
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        match_count:
          type: integer
          title: Match Count
          default: 0
      type: object
      required:
      - name
      - terms
      - id
      title: WatchlistResponse
    WatchlistUpdate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        terms:
          items:
            type: string
          type: array
          maxItems: 1000
          minItems: 1
          title: Terms
        notify:
          type: boolean
          title: Notify
          default: true
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - name
      - terms
      title: WatchlistUpdate
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8003